fix: unbloat

+++ asm/hash.c
@@ -15,7 +15,6 @@
                                       |___/
 */
 
-#include <sys/types.h>
 #include "asm.h"
 
 unsigned long hash(char *str)
@@ -23,7 +22,7 @@ unsigned long hash(char *str)
     unsigned long hash = 5381;
     int c;
     while ((c = (unsigned char)*str++))
-        hash = ((hash << 5) + hash) + (ulong)c;
+        hash = ((hash << 5) + hash) + (size_t)c;
     return hash % TABLE_SIZE;
 }
